+++ src/api/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path, re_path
 from django.conf.urls import url
-from .views_new import *
+from .views import *
 
 urlpatterns = [
   url(r'^$', ping),
